update to gcr.io/istio-testing versions (#170)
1. Update mixer, manager, proxy versions to include rate limit fixes
2. Remove mixer configmap. The default config is now baked inside mixer.
3. expose mixer metrics and configapi ports thru port forwarding.
4. Add "wrk" for testing. drive traffic and fetch metrics as a setup for full
5. ratelimit integration test. That PR will follow.

+++ tests/kubeUtils.sh
@@ -48,6 +48,7 @@ function find_istio_endpoints() {
       -o jsonpath='{.items[*].subsets[*].addresses[*].ip}'))
     echo ${endpoints[@]}
     [[ ${#endpoints[@]} -eq 4 ]] && return 0
+    ${K8CLI} get endpoints -n ${NAMESPACE}
     return 1
 }
 
@@ -63,8 +64,24 @@ function setup_istioctl(){
 # Kill the port forwarding process
 function cleanup_istioctl(){
     print_block_echo "Cleaning up istioctl"
-    kill ${pfPID}
+    if [[ ! -z ${pfPID} ]];then
+      kill ${pfPID}
+    fi
+    if [[ ! -z ${pfPID2} ]];then
+      kill ${pfPID2}
+    fi
 }
+
+# Port forward mixer
+function setup_mixer(){
+    print_block_echo "Setting up mixer"
+    ${K8CLI} -n ${NAMESPACE} port-forward $(${K8CLI} -n ${NAMESPACE} get pod -l istio=mixer \
+    -o jsonpath='{.items[0].metadata.name}') 9091 9094 9095:42422 &
+    pfPID2=$!
+    export ISTIO_MIXER_METRICS=http://localhost:9095
+    export ISTIO_MIXER_CONFIGAPI=http://localhost:9094
+}
+
 
 # Deploy the bookinfo microservices
 function deploy_bookinfo() {
@@ -77,18 +94,21 @@ function deploy_bookinfo() {
 }
 
 function find_ingress_controller() {
-    #local gateway="$(${K8CLI} get svc istio-ingress -n ${NAMESPACE} \
-    #  -o jsonpath='{.status.loadBalancer.ingress[*].ip}')"
-    #if [[ ${gateway} =~ ^[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}$ ]]; then
-    #    GATEWAY_URL="http://${gateway}"
-    #    return 0
-    #fi
-    local gateway="$(${K8CLI} get po -l istio=ingress -n ${NAMESPACE} \
-      -o jsonpath='{.items[0].status.hostIP}'):$(${K8CLI} get svc istio-ingress -n ${NAMESPACE} \
-      -o jsonpath={.spec.ports[0].nodePort})"
-    if [[ ${gateway} =~ ^[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\:3[0-2][0-9][0-9][0-9]$ ]]; then
-        GATEWAY_URL="http://${gateway}"
-        return 0
+    if [[ ! -z ${OUT_OF_CLUSTER} ]]; then
+      local gateway="$(${K8CLI} get svc istio-ingress -n ${NAMESPACE} \
+        -o jsonpath='{.status.loadBalancer.ingress[0].ip}')"
+      if [[ ${gateway} =~ ^[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}$ ]]; then
+          GATEWAY_URL="http://${gateway}"
+          return 0
+      fi
+    else
+      local gateway="$(${K8CLI} get po -l istio=ingress -n ${NAMESPACE} \
+        -o jsonpath='{.items[0].status.hostIP}'):$(${K8CLI} get svc istio-ingress -n ${NAMESPACE} \
+        -o jsonpath={.spec.ports[0].nodePort})"
+      if [[ ${gateway} =~ ^[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\:3[0-2][0-9][0-9][0-9]$ ]]; then
+          GATEWAY_URL="http://${gateway}"
+          return 0
+      fi
     fi
     return 1
 }
@@ -106,7 +126,7 @@ function cleanup() {
 # Debug dump for failures
 function dump_debug() {
     echo ""
-    $K8CLI -n $NAMESPACE get pods
+    $K8CLI -n $NAMESPACE get pods,deployments,svc,endpoints
     $K8CLI -n $NAMESPACE get thirdpartyresources
     $K8CLI -n $NAMESPACE get thirdpartyresources -o json
     GATEWAY_PODNAME=$($K8CLI -n $NAMESPACE get pods | grep istio-ingress | awk '{print $1}')
